backendv2/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.21.1
httpx==0.24.1